Build from submodules
Builds most included projects from submodules. Citra and devkitARM are
still downloaded to avoid setting up a secondary build environment.
The missing citro3d submodule has been added. There are so many
homebrew apps with a c that I kinda forgot it.

The urls for citra and devkitARM are now defined in the build script.
A typo with 3ds-portlib's tar-source has been fixed. To avoid having
to send 500MB to the docker daemon the Dockerfile and imagefs are
copied into the build directory alongside the compressed sources.

The sources are now copied instead of the whole project root to avoid
copying the `.git` directories which would bloat the image. Since
in-image downloads are now almost completely removed most ARGs could
be deleted and default values were removed to force usage of build.sh.
Some ENVs were removed because they were unused or set by the system.

The usage of git is disabled for 3ds-portlibs since all .git folders
are removed. It is useless, but I'm too lazy to change that now.
mbedtls-apache has become the only option in my fork and thus been
renamed to mbedtls.

Citra was now moved to the previously used version. I think I forgot
changing it from the latest commit when I originally added it.